Add copyconfig.
With copyconfig it is possible to retrieve the configuration
from an existing LKFT build from Jenkins.

This takes an URL in the like the following:
  https://ci.linaro.org/job/openembedded-lkft-linux-mainline/818/DISTRO=rpb,MACHINE=hikey,label=docker-stretch-amd64/
or:
  https://ci.linaro.org/job/openembedded-lkft-linux-mainline/DISTRO=rpb,MACHINE=hikey,label=docker-stretch-amd64/818/

This is a work in progress that needs better handling of the
URL, like validation and proper linking between snapshots and
build job, but at least it works for now.

Signed-off-by: Daniel Díaz <daniel.diaz@linaro.org>